chore: move commands to common and cross referfence in cli also
